Update typings test setup
Use dtslint for testing, which is a bit more stable. Test on the
.d.ts files from the library output to make sure we assert on the
same typings as the library consumers
