+++ package.json
@@ -14,7 +14,7 @@
     "build:babel": "babel ./src -x \".ts\" -x \".js\" --out-dir ./",
     "build:ts": "tsc -p ./tsconfig.build.json && shx cp -Rf decl/* .",
     "test": "jest",
-    "test:types": "typings-checker --allow-expect-error --project tsconfig.json test/types/tests.ts",
+    "test:types": "npm run build:ts && dtslint test/types",
     "test:dev": "jest --watch",
     "clean": "npm-run-all clean:*",
     "clean:test": "shx rm -rf coverage",
@@ -75,6 +75,7 @@
     "babel-eslint": "^8.0.3",
     "coveralls": "^3.0.2",
     "cross-env": "^5.1.1",
+    "dtslint": "^0.4.9",
     "eslint": "^4.13.1",
     "eslint-config-airbnb-base": "^12.1.0",
     "eslint-config-prettier": "^2.9.0",
@@ -91,8 +92,7 @@
     "tslint-config-airbnb": "^5.11.1",
     "tslint-config-prettier": "^1.18.0",
     "typedoc": "^0.9.0",
-    "typescript": "^3.3.3",
-    "typings-checker": "^2.0.0"
+    "typescript": "^3.3.3"
   },
   "dependencies": {
     "@babel/runtime": "^7.3.1",
