+++ test/types/index.d.ts
@@ -1,0 +1,4 @@
+// TypeScript Version: 3.1
+
+// tslint:disable-next-line
+/// <reference path="../../index.d.ts" />
