+++ test/types/tsconfig.json
@@ -1,0 +1,18 @@
+{
+  "compilerOptions": {
+    "module": "es2015",
+    "lib": ["DOM", "DOM.Iterable", "ES2017"],
+    "moduleResolution": "node",
+    "noImplicitAny": true,
+    "noImplicitThis": true,
+    "strictNullChecks": true,
+    "strictFunctionTypes": true,
+    "noEmit": true,
+    "noErrorTruncation": true,
+    "baseUrl": "../../",
+    "skipLibCheck": true,
+    "paths": {
+      "seng-event": ["."]
+    }
+  }
+}
