+++ tsconfig.json
@@ -14,6 +14,7 @@
     "noUnusedLocals": true,
     "noUnusedParameters": true,
     "strictNullChecks": true,
+    "noErrorTruncation": true,
     "pretty": true,
     "skipLibCheck": true,
     "sourceMap": true,
